+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gumbee/agent
 
+## 1.1.0
+
+### Minor Changes
+
+- added TokenWindowMemory, SlidingWindowMemory
+- switched agent loop steps to be 0-based instead of 1-based
+- removed hard limit on agent loop counts (can execute indefinitely). Stop conditions are now the sole origin of stopping
+
 ## 1.0.0
 
 ### Major Changes